docs: document query for creating mock accounts
Sometimes when pulling the ACC database for local testing having mock
accounts may be desired. These may or may no longer exist on ACC which
makes the mock-login service fail. Figuring out which query to execute
to have working mock accounts can be annoying, so this should make it
easy to copy paste.